Update HTML tags to include IDs for testing dynamic content - sufrace… (#117)
* Update HTML tags to include IDs for testing dynamic content - sufrace water

https://eaflood.atlassian.net/browse/LTFRI-884

Following a meeting with Chris, he advised that the tests are more complicated than they could be. This is due to the dynamic content not have unique references to narrow down the tests. He has suggest we add IDs to the dynamic content in order to increase reliability of the tests.