Code Scenarios selection for new map design (#132)
https://eaflood.atlassian.net/browse/LTFRI-781

With the new map there will be map layout options that will trigger a scenario key to appear at the bottom of the map. This will give the user ways to alter the map layers.

https://eaflood.atlassian.net/browse/LTFRI-860

Users visiting the map and selecting the advanced options link they should see the reservoirs extent results on the map.