Map users from the postcode page link
https://eaflood.atlassian.net/browse/LTFRI-795

The map that will be accessible from the postcode page will be full screen, there will be no address or boundary pointers. The user will navigate to their desired location by dragging/scrolling around the map. The full key should be visible in this version of the map with no simple/advanced toggle. The user will be able to return to the postcode screen from this map.