Surface water and Rivers and Sea options for holding comments
https://eaflood.atlassian.net/browse/LTFRI-1020

The holding comments form was designed to support changes to surface water data. The option to override SW data is the first option on the holding comments form. As we’re now using this for adding a free text option for rivers and sea, it will be confusing for users to see this option first. We want to simplify this for users while adding in the option for Rivers and the sea to be included.